+++ src/main/java/entity/User.java
@@ -1,4 +1,4 @@
-package entities;
+package entity;
 
 import java.util.List;
